requirements.txt
sanic==0.5.4
Jinja2==2.9.6
sanic-jinja2==0.5.1
peewee-async==0.5.7
aiomysql==0.0.9
python-paginate==0.3.5